requirements.txt
django==4
djangorestframework==3.13.1
psycopg2-binary==2.8.6
# django-silk==4.2.0
locust==2.5.1
gunicorn==20.1.0

